app/requirements.txt
langchain==0.3.27
langchain_community==0.3.29
langchain-openai==0.3.32
faiss-cpu==1.12.0
pypdf==6.0.0
flask==3.1.2
python-dotenv==1.1.1